requirements.txt
Flask==3.1.0
flask-rdf==0.2.1
lxml==5.4.0
rdflib==7.1.4
SPARQLWrapper==2.0.0
flask_bootstrap
flask_sqlalchemy
flask_wtf
flask_bcrypt
flask_login
flask_mysql
wtforms
email_validator
python-dotenv